requirement.txt
Flask==3.1.1
python-dotenv==1.0.0
google-generativeai==0.3.2
firebase-admin==6.2.0
gunicorn==21.2.0
Werkzeug==3.0.1